agentic_bug_hunter/requirements.txt
pandas
requests
tk
python-dotenv
flask
flask-cors